rename: Item 컴포넌트 위치 변경
shop과 decorate 페이지에서 공동으로 사용되지만 따로 구분되던  Item 컴포넌트를 하나의 컴포넌트로 합쳤습니다.
